# [devsync] add cudaDeviceSynchronize for better timing measurements