--- a/epoch1/cuda/ee_mumu/SubProcesses/P1_Sigma_sm_epem_mupmum/throughput12.sh
+++ b/epoch1/cuda/ee_mumu/SubProcesses/P1_Sigma_sm_epem_mupmum/throughput12.sh
@@ -1,14 +1,18 @@
 #!/bin/bash
 
+set +x
+
 omp=0
 avxall=0
 ep2=0
 cpp=1
 ab3=0
+ggttgg=0
+verbose=0
 
 function usage()
 {
-  echo "Usage: $0 [-nocpp|[-omp][-avxall]] [-ep2] [-3a3b]"
+  echo "Usage: $0 [-nocpp|[-omp][-avxall]] [-ep2] [-3a3b] [-ggttgg] [-v]"
   exit 1
 }
 
@@ -32,6 +36,12 @@ while [ "$1" != "" ]; do
   elif [ "$1" == "-3a3b" ]; then
     ab3=1
     shift
+  elif [ "$1" == "-ggttgg" ]; then
+    ggttgg=1
+    shift
+  elif [ "$1" == "-v" ]; then
+    verbose=1
+    shift
   else
     usage
   fi
@@ -39,17 +49,17 @@ done
 
 exes=
 
-#===============================
-# CUDA (epoch1 and epoch2)
-#===============================
+#=====================================
+# CUDA (eemumu/epoch1, eemumu/epoch2)
+#=====================================
 exes="$exes ../../../../../epoch1/cuda/ee_mumu/SubProcesses/P1_Sigma_sm_epem_mupmum/build.none/gcheck.exe"
 if [ "${ep2}" == "1" ]; then 
   exes="$exes ../../../../../epoch2/cuda/ee_mumu/SubProcesses/P1_Sigma_sm_epem_mupmum/gcheck.exe"
 fi
 
-#===============================
-# C++ (epoch1 and epoch2)
-#===============================
+#=====================================
+# C++ (eemumu/epoch1, eemumu/epoch2)
+#=====================================
 if [ "${cpp}" == "1" ]; then 
   exes="$exes ../../../../../epoch1/cuda/ee_mumu/SubProcesses/P1_Sigma_sm_epem_mupmum/build.none/check.exe"
 fi
@@ -66,6 +76,22 @@ fi
 if [ "${ep2}" == "1" ]; then 
   if [ "${cpp}" == "1" ]; then 
     exes="$exes ../../../../../epoch2/cuda/ee_mumu/SubProcesses/P1_Sigma_sm_epem_mupmum/check.exe"
+  fi
+fi
+
+#=====================================
+# CUDA (ggttgg/epoch2)
+#=====================================
+if [ "${ggttgg}" == "1" ]; then 
+  exes="$exes ../../../../../epoch2/cuda/gg_ttgg/SubProcesses/P1_Sigma_sm_gg_ttxgg/gcheck.exe"
+fi
+
+#=====================================
+# C++ (ggttgg/epoch2)
+#=====================================
+if [ "${ggttgg}" == "1" ]; then 
+  if [ "${cpp}" == "1" ]; then 
+    exes="$exes ../../../../../epoch2/cuda/gg_ttgg/SubProcesses/P1_Sigma_sm_gg_ttxgg/check.exe"
   fi
 fi
 
@@ -86,8 +112,16 @@ if [ "${ep2}" == "1" ]; then
   popd >& /dev/null
 fi
 
+if [ "${ggttgg}" == "1" ]; then 
+  pushd ../../../../../epoch2/cuda/gg_ttgg/SubProcesses/P1_Sigma_sm_gg_ttxgg >& /dev/null
+  pwd
+  make
+  popd >& /dev/null
+fi
+
 function runExe() {
   exe=$1
+  args="$2"
   ###echo "runExe $exe OMP=$OMP_NUM_THREADS"
   pattern="Process|fptype_sv|OMP threads|EvtsPerSec\[Matrix|MeanMatrix|FP precision|TOTAL       :"
   # Optionally add other patterns here for some specific configurations (e.g. clang)
@@ -98,24 +132,39 @@ function runExe() {
     # -- Newer version using perf stat
     pattern="${pattern}|instructions|cycles"
     pattern="${pattern}|elapsed"
-    perf stat $exe -p 2048 256 12 2>&1 | egrep "(${pattern})" | grep -v "Performance counter stats"
+    if [ "${verbose}" == "1" ]; then set -x; fi
+    perf stat $exe $args 2>&1 | egrep "(${pattern})" | grep -v "Performance counter stats"
+    set +x
   else
     # -- Older version using time
     # For TIMEFORMAT see https://www.gnu.org/software/bash/manual/html_node/Bash-Variables.html
-    TIMEFORMAT=$'real\t%3lR' && time $exe -p 2048 256 12 2>&1 | egrep "(${pattern})"
+    if [ "${verbose}" == "1" ]; then set -x; fi
+    TIMEFORMAT=$'real\t%3lR' && time $exe $args 2>&1 | egrep "(${pattern})"
+    set +x
   fi
 }
 
 function runNcu() {
   exe=$1
+  args="$2"
   ###echo "runExe $exe OMP=$OM_NUM_THREADS (NCU)"
-  $(which ncu) --metrics launch__registers_per_thread --target-processes all --kernel-id "::sigmaKin:" --print-kernel-base mangled $exe -p 2048 256 1 | egrep '(sigmaKin|registers)' | tr "\n" " " | awk '{print $1, $2, $3, $15, $17}'
+  if [ "${verbose}" == "1" ]; then set -x; fi
+  $(which ncu) --metrics launch__registers_per_thread --target-processes all --kernel-id "::sigmaKin:" --print-kernel-base mangled $exe $args | egrep '(sigmaKin|registers)' | tr "\n" " " | awk '{print $1, $2, $3, $15, $17}'
+  set +x
 }
 
 lastExe=
 echo -e "\nOn $HOSTNAME ($(nvidia-smi -L | awk '{print $5}')):"
 for exe in $exes; do
   if [ ! -f $exe ]; then continue; fi
+  if [ "${exe%%/gg_ttgg*}" != "${exe}" ]; then 
+    # This is a good GPU middle point: tput is 1.5x lower with "32 256 1", only a few% higher with "128 256 1"
+    exeArgs="-p 64 256 1"
+    ncuArgs="-p 64 256 1"
+  else
+    exeArgs="-p 2048 256 12"
+    ncuArgs="-p 2048 256 1"
+  fi
   if [ "$(basename $exe)" != "$lastExe" ]; then
     echo "========================================================================="
     lastExe=$(basename $exe)
@@ -123,16 +172,16 @@ for exe in $exes; do
     echo "-------------------------------------------------------------------------"
   fi
   unset OMP_NUM_THREADS
-  runExe $exe
+  runExe $exe "$exeArgs"
   if [ "${exe%%/check*}" != "${exe}" ]; then 
     obj=${exe%%/check*}/CPPProcess.o; ./simdSymSummary.sh -stripdir ${obj}
     if [ "${omp}" == "1" ]; then 
       echo "-------------------------------------------------------------------------"
       export OMP_NUM_THREADS=$(nproc --all)
-      runExe $exe
+      runExe $exe "$exeArgs"
     fi
   elif [ "${exe%%/gcheck*}" != "${exe}" ]; then 
-    runNcu $exe
+    runNcu $exe "$ncuArgs"
   fi
 done
 echo "========================================================================="