@roiser @oliviermattelaer this is related to what we discussed during the last week

I am adding a cudaDeviceSynchronize between sigmakin and the copy of ME from device to host

The overall throughput does not change (*), while the relative split of sigmakin and copyDtoH changes enormously. Now it is clear that the copy takes around the same time as the sigmakin calculation for eemumu, while for ggttgg it is totally negligible with respect to it.

(*) NB: I am observing big fluctuations on the VM, and they actually come from the copy, while the calculation is rather stable. For eemumu this transaltes into large fluctuations (between 6.0 and 6.5E8, lower than the usual 7.3E8 that it should be). This is discussed in https://cern.service-now.com/service-portal?id=ticket&table=u_request_fulfillment&n=RQF1789593

Before adding cudaDeviceSynchronzie
```
On itscrd70.cern.ch (V100S-PCIE-32GB):
=========================================================================
Process                     = EPOCH1_EEMUMU_CUDA [nvcc 11.0.221]
FP precision                = DOUBLE (NaN/abnormal=0, zero=0)
EvtsPerSec[MatrixElems] (3) = ( 6.568678e+08                 )  sec^-1
MeanMatrixElemValue         = ( 1.371706e-02 +- 3.270315e-06 )  GeV^0
3a SigmaKin :     0.000086 sec
3b CpDTHmes :     0.009492 sec
TOTAL       :     0.742937 sec
     2,594,268,110      cycles                    #    2.658 GHz
     3,534,129,010      instructions              #    1.36  insn per cycle
       1.045660455 seconds time elapsed
==PROF== Profiling "sigmaKin": launch__registers_per_thread 120
-------------------------------------------------------------------------
FP precision               = DOUBLE (nan=0)
EvtsPerSec[MatrixElems] (3)= ( 4.397884e+05                 )  sec^-1
MeanMatrixElemValue        = ( 5.532387e+01 +- 5.501866e+01 )  GeV^-4
3a SigmaKin :     0.000009 sec
3b CpDTHmes :     0.037245 sec
TOTAL       :     0.606138 sec
     2,201,855,033      cycles                    #    2.656 GHz
     2,955,778,217      instructions              #    1.34  insn per cycle
       0.890392160 seconds time elapsed
==PROF== Profiling "sigmaKin": launch__registers_per_thread 255
=========================================================================
``` 

After adding cudaDeviceSynchronize
```
On itscrd70.cern.ch (V100S-PCIE-32GB):
=========================================================================
Process                     = EPOCH1_EEMUMU_CUDA [nvcc 11.0.221]
FP precision                = DOUBLE (NaN/abnormal=0, zero=0)
EvtsPerSec[MatrixElems] (3) = ( 6.480230e+08                 )  sec^-1
MeanMatrixElemValue         = ( 1.371706e-02 +- 3.270315e-06 )  GeV^0
3a SigmaKin :     0.004613 sec
3b CpDTHmes :     0.005095 sec
TOTAL       :     0.743895 sec
     2,601,684,149      cycles                    #    2.657 GHz
     3,532,142,253      instructions              #    1.36  insn per cycle
       1.049697294 seconds time elapsed
==PROF== Profiling "sigmaKin": launch__registers_per_thread 120
-------------------------------------------------------------------------
FP precision               = DOUBLE (nan=0)
EvtsPerSec[MatrixElems] (3)= ( 4.427943e+05                 )  sec^-1
MeanMatrixElemValue        = ( 5.532387e+01 +- 5.501866e+01 )  GeV^-4
3a SigmaKin :     0.036976 sec
3b CpDTHmes :     0.000026 sec
TOTAL       :     0.651707 sec
     2,224,902,887      cycles                    #    2.657 GHz
     2,976,495,038      instructions              #    1.34  insn per cycle
       0.941813933 seconds time elapsed
==PROF== Profiling "sigmaKin": launch__registers_per_thread 255
=========================================================================
```

By the way, I am using 2048 256 12 for eemumu, against 64 256 1 for ggttgg. This means I am copying 32x12 ie 384 times less data in ggttgg. In eemumu it takes 5095 usec, in ggttgg 26 usec to copy the data, so a factor 200 difference, this looks vaguely reasonable.